refactor: Extract utilities and download logic from MFARecoveryCodes component
