+++ src/modules/Dashboard/features/AccountSettings/MFARecoveryCodes/index.ts
@@ -1,2 +1,4 @@
 export { MFARecoveryCodes } from './MFARecoveryCodes';
 export type { MFARecoveryCodesProps } from './MFARecoveryCodes.types';
+export { useRecoveryCodesDownload } from './useRecoveryCodesDownload';
+export * from './MFARecoveryCodes.utils';
